[Fix] Update references change ids to JPA entitites, some breaking type changes from Long to Integer.

+++ src/main/java/com/ormi5/movieblog/post/PostDto.java
@@ -20,7 +20,7 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 @Builder
 public class PostDto implements Serializable {
-    Long postId;
+    Integer postId;
     UserDto user;
     String title;
     String content;
